Consolidate transformation case studies
- Remove transformation, transformation1, and transformation2 folders
- Rename transformation3 to transformation as the final version
